requirements.txt
Django==5.2.3
djangorestframework==3.15.1
python-decouple==3.8
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
